+++ src/cpu/tms34010_intf.h
@@ -39,7 +39,7 @@ UINT16 TMS34010ReadWord(UINT32 address);
 void TMS34010WriteWord(UINT32 address, UINT16 value);
 void TMS34010MapReset();
 void TMS34010MapMemory(UINT8 *mem, UINT32 start, UINT32 end, UINT8 type);
-void TMS34010MapHandler(UINT32 num, UINT32 start, UINT32 end, UINT8 type);
+void TMS34010MapHandler(uintptr_t num, UINT32 start, UINT32 end, UINT8 type);
 int TMS34010SetReadHandler(UINT32 num, pTMS34010ReadHandler handler);
 int TMS34010SetWriteHandler(UINT32 num, pTMS34010WriteHandler handler);
 int TMS34010SetHandlers(UINT32 num, pTMS34010ReadHandler rhandler, pTMS34010WriteHandler whandler);
